input: move input code into the input/ directory
And this make sure the input code is self-contained.
