+++ src/input/input.c
@@ -38,12 +38,12 @@
 #include <string.h>
 #include <unistd.h>
 #include "eloop.h"
+#include "input.h"
+#include "input_internal.h"
 #include "shl_dlist.h"
 #include "shl_hook.h"
 #include "shl_log.h"
 #include "shl_misc.h"
-#include "uterm_input.h"
-#include "uterm_input_internal.h"
 
 #define LOG_SUBSYSTEM "uterm_input"
