+++ src/input/input_pointer.c
@@ -2,9 +2,9 @@
 #include <linux/input.h>
 #include <time.h>
 #include "eloop.h"
+#include "input.h"
+#include "input_internal.h"
 #include "shl_hook.h"
-#include "uterm_input.h"
-#include "uterm_input_internal.h"
 
 static void pointer_update_inactivity_timer(struct uterm_input_dev *dev)
 {
