+++ src/kmscon_main.c
@@ -33,6 +33,7 @@
 #include <sys/signalfd.h>
 #include "conf.h"
 #include "eloop.h"
+#include "input/input.h"
 #include "kmscon_conf.h"
 #include "kmscon_seat.h"
 #include "shl_dlist.h"
@@ -40,7 +41,6 @@
 #include "shl_misc.h"
 #include "shl_module.h"
 #include "text.h"
-#include "uterm_input.h"
 #include "uterm_monitor.h"
 #include "uterm_video.h"
 #include "uterm_vt.h"
