+++ src/uterm_vt.h
@@ -33,10 +33,10 @@
 #define UTERM_UTERM_VT_H
 
 #include <eloop.h>
+#include <input/input.h>
 #include <inttypes.h>
 #include <stdbool.h>
 #include <stdlib.h>
-#include <uterm_input.h>
 
 struct uterm_vt;
 struct uterm_vt_master;
