+++ src/uterm_vt_real.c
@@ -67,10 +67,10 @@
 #include <unistd.h>
 
 #include "eloop.h"
+#include "input/input.h"
+#include "input/input_internal.h"
 #include "shl_log.h"
 #include "shl_misc.h"
-#include "uterm_input.h"
-#include "uterm_input_internal.h"
 #include "uterm_vt.h"
 #include "uterm_vt_internal.h"
