+++ tests/test_vt.c
@@ -43,9 +43,9 @@ static void print_help();
 #include <string.h>
 #include <unistd.h>
 #include "eloop.h"
+#include "input/input.h"
 #include "shl_log.h"
 #include "test_include.h"
-#include "uterm_input.h"
 #include "uterm_vt.h"
 
 static void print_help()
